refactor: simplify source_search slice.
