--- lib/features/source_search/domain/domain.dart
@@ -1,1 +1,0 @@
-export 'services/fetch_search.dart';
