+++ lib/features/source_search/model/search_info.dart
