+++ lib/features/source_search/source_search.dart
@@ -1,8 +1,8 @@
-export 'presentation/model/search_info.dart';
+export 'model/search_info.dart';
 
-export 'presentation/provider/crawler_search_provider.dart';
-export 'presentation/provider/is_searching_provider.dart';
-export 'presentation/provider/search_fetch_provider.dart';
-export 'presentation/provider/search_provider.dart';
+export 'provider/crawler_search_provider.dart';
+export 'provider/is_searching_provider.dart';
+export 'provider/search_fetch_provider.dart';
+export 'provider/search_provider.dart';
 
-export 'presentation/widgets/search_button.dart';
+export 'widgets/search_button.dart';
